Harmonogram do Kalkulatora

+++ src/main/java/Calculator/Coderion/Calculator/CalculatorService.java
@@ -1,20 +1,20 @@
 package Calculator.Coderion.Calculator;
+
 import java.math.BigDecimal;
 import java.math.RoundingMode;
 
 public class CalculatorService {
     public BigDecimal rataKredytu;
 
-    public BigDecimal calculate(int kwota, int liczba_rat,int suma_odsetek,int Rata_kapitalowa_techniczna) {
+    public BigDecimal calculate(int kwota, int liczba_rat, int suma_odsetek, double Rata_kapitalowa_techniczna) {
         int oprocentowanie_roczne = 23;
-        Rata_kapitalowa_techniczna = (int) kwota /liczba_rat ;
+        Rata_kapitalowa_techniczna = (double) kwota / liczba_rat;
         double oprocentowanie_miesieczne = (double) oprocentowanie_roczne / 12 / 100;
-        double rata_kredytu = (double) (kwota + suma_odsetek) /liczba_rat;
+        double rata_kredytu = (double) (kwota + suma_odsetek) / liczba_rat;
+
         rataKredytu = BigDecimal.valueOf(rata_kredytu).setScale(2, RoundingMode.HALF_UP);
         return rataKredytu;
     }
 
 
-
 }
-
